Rename lightning module name
