+++ convmodel/lightning_model.py
