+++ convmodel/trainer.py
@@ -1,4 +1,4 @@
 from pytorch_lightning.utilities.cli import LightningCLI
-from .pl_model import LightningConversationModel
+from .lightning_model import LightningConversationModel
 
 cli = LightningCLI(LightningConversationModel)
